Modify messages to deal with unaccounted changes
Further modified messages to comply with reduced requirements.
Specifically, to deal with changes that were not accounted for in the
previous commit, such as a couple of schema changes. Still, currently
untested due to lack of mock data and inability to request coaches and
accept clients due to the related endpoints having been updated.