Comment out broken and unused function
Commented out a series of functions dealing with getting a user profile
by ID. This endpoint and its related functions are never used and
when I tried to test them, they were broken. I fixed them up so it was
at least not crashing, but it's missing a lot of features and checks,
like validation and ensuring only authorized users see particular
profiles. I will leave them in the code base just in case someone's
is interested in completing them.